refactor; make it work
